Move control of connections from NetworkController to P2P. "Testing Ping-Pong Fail" test will fail until accepting connection logic refactored.
