+++ net/src/main/java/io/bitcoinsv/bsvcl/net/P2P.java
@@ -1,11 +1,15 @@
 package io.bitcoinsv.bsvcl.net;
 
 
+import com.google.common.util.concurrent.Service;
+import io.bitcoinsv.bsvcl.common.thread.TimeoutTask;
+import io.bitcoinsv.bsvcl.common.thread.TimeoutTaskBuilder;
+import io.bitcoinsv.bsvcl.net.network.events.*;
+import io.bitcoinsv.bsvcl.net.network.streams.nio.NIOStream;
 import io.bitcoinsv.bsvcl.net.protocol.handlers.blacklist.BlacklistHandler;
 import io.bitcoinsv.bsvcl.net.protocol.handlers.blacklist.BlacklistHandlerImpl;
 import io.bitcoinsv.bsvcl.net.protocol.handlers.handshake.HandshakeHandler;
 import io.bitcoinsv.bsvcl.net.network.PeerAddress;
-import io.bitcoinsv.bsvcl.net.network.events.HandlerStateEvent;
 import io.bitcoinsv.bsvcl.net.network.NetworkController;
 import io.bitcoinsv.bsvcl.net.protocol.config.ProtocolConfig;
 import io.bitcoinsv.bsvcl.net.protocol.config.provided.ProtocolBSVMainConfig;
@@ -19,10 +23,14 @@ import io.bitcoinsv.bsvcl.net.tools.LoggerUtil;
 import io.bitcoinsv.bsvcl.common.handlers.HandlerConfig;
 import io.bitcoinsv.bsvcl.common.thread.ThreadUtils;
 
+import java.net.InetAddress;
 import java.net.UnknownHostException;
+import java.nio.channels.SelectionKey;
 import java.time.Duration;
 import java.util.*;
 import java.util.concurrent.*;
+import java.util.function.Predicate;
+import java.util.stream.Collectors;
 
 /**
  * The P2P class is the controller for accessing the Bitcoin P2P network.
@@ -78,12 +86,26 @@ public class P2P extends Thread {
 
     // Latch that is released when start has finished.
     private final CountDownLatch startedLatch = new CountDownLatch(1);
-    // Latch that is released when P2P should stop
+    // Latch to indicate that the P2P should shut down
     private final CountDownLatch stopLatch = new CountDownLatch(1);
+    // Semaphore that is incremented when there are events to process.
+    private final Semaphore eventsToProcess = new Semaphore(1);
+    // The active connections. Note that this class considers a connection active as soon as a NetworkController
+    // has been instructed to open it.
+    private final Map<PeerAddress, NetworkController> activeConns = new ConcurrentHashMap<>();
+    // the connections that are waiting to be opened
+    private final BlockingQueue<PeerAddress> pendingToOpenConns = new LinkedBlockingQueue<>();
+    // the connections that are waiting to be closed
+    private final BlockingQueue<DisconnectPeerRequest> pendingToCloseConns = new LinkedBlockingQueue<>();
+    // the addresses which have been tried and are now closed
+    private final Set<PeerAddress> closedConns = ConcurrentHashMap.newKeySet();
+    // the addresses which are blacklisted
+    private final Set<InetAddress> blacklist = ConcurrentHashMap.newKeySet();
+    // the addresses which have failed to connect
+    private final Set<PeerAddress> failedConns = ConcurrentHashMap.newKeySet();
 
     // Event Stream Managers Definition:
     public final P2PEventStreamer EVENTS;
-
     // Request Handler:
     public final P2PRequestHandler REQUESTS;
 
@@ -131,8 +153,78 @@ public class P2P extends Thread {
         this.stateRefreshFrequencies.put(handlerId, frequency);
     }
 
+    /** Add a Peer to the list of Peers to connect to. */
+    public void connect(PeerAddress peerAddress) {
+        connect(Collections.singletonList(peerAddress));
+    }
+
+    /** Add a list of Peers to the list of Peers to connect to. */
+    public void connect(List<PeerAddress> peerAddresses) {
+        if (peerAddresses == null) return;
+        // remove the Peers we are already connected to, or in process to...
+        List<PeerAddress> listToAdd = peerAddresses.stream()
+                .filter(p -> !activeConns.containsKey(p))
+                .filter(p -> !pendingToOpenConns.contains(p))
+                .filter(p -> !pendingToCloseConns.contains( new PeerAddress2DisconnectPeerRequest_Comparator(p) ))
+                .filter(p -> !blacklist.contains(p.getIp()))
+                .collect(Collectors.toList());
+
+        if (listToAdd.size() > 0) {
+            // Check that the limit in the Pending Socket Connections is not exceeded
+            // If there is no limit, we just include them all. If there is a limit, we only include them up
+            // to the limit.
+            List<PeerAddress> finalListToAdd = listToAdd;
+
+            int limit = networkConfig.getMaxSocketPendingConnections();
+            int numItemsToAdd = Math.min(finalListToAdd.size(), limit - pendingToOpenConns.size());
+            if (numItemsToAdd > 0)
+                finalListToAdd = listToAdd.subList(0, numItemsToAdd);
+            else finalListToAdd = new ArrayList<>(); // empty List
+            pendingToOpenConns.addAll(finalListToAdd);
+            eventsToProcess.release();
+        }
+    }
+
+    public void disconnect(PeerAddress peerAddress) {
+        processDisconnectRequest(new DisconnectPeerRequest(peerAddress, PeerDisconnectedEvent.DisconnectedReason.DISCONNECTED_BY_LOCAL));
+    }
+
+    public void disconnect(List<PeerAddress> peerAddressList) {
+        processDisconnectRequests(peerAddressList
+                .stream()
+                .map(p -> new DisconnectPeerRequest(p, PeerDisconnectedEvent.DisconnectedReason.DISCONNECTED_BY_LOCAL))
+                .collect(Collectors.toList()));
+    }
+
+    public void blacklist(InetAddress ipAddress, PeersBlacklistedEvent.BlacklistReason reason) {
+        Map<InetAddress, PeersBlacklistedEvent.BlacklistReason> map = new HashMap<>();
+        map.put(ipAddress, reason);
+        blacklist(map);
+    }
+
+    public void blacklist(Map<InetAddress, PeersBlacklistedEvent.BlacklistReason> ipAddresses) {
+        if (ipAddresses == null) return;
+        // First, we add the IpAddress to the Blacklist, to keep a reference to them.
+        blacklist.addAll(ipAddresses.keySet());
+
+        // Then, we disconnect all the current Peers already connected to any of those addresses...
+        List<DisconnectPeerRequest> requestsToDisconnect = this.activeConns.keySet().stream()
+                .filter(p -> ipAddresses.keySet().contains(p.getIp()))
+                .map(p -> new DisconnectPeerRequest(p, PeerDisconnectedEvent.DisconnectedReason.DISCONNECTED_BY_LOCAL_BLACKLIST))
+                .collect(Collectors.toList());
+        this.processDisconnectRequests(requestsToDisconnect);
+    }
+
+    public void removeFromBlacklist(List<InetAddress> ipAddresses) {
+        if (ipAddresses == null) return;
+        ipAddresses.forEach(blacklist::remove);
+    }
+
     public void run() {
         try {
+            // register for events first, so we don't miss any
+            registerForEvents();
+
             // set up the network controllers, at the moment we only have one
             String serverIp = "0.0.0.0:" + networkConfig.getListeningPort();
             this.networkController = new NetworkController(id, runtimeConfig, networkConfig, PeerAddress.fromIp(serverIp), networkConfig.isListening());
@@ -142,7 +234,6 @@ public class P2P extends Thread {
             // If specified, we trigger a new Thread that will publish the status of the Handlers into the
             // Bus. The Map contains a duration for each Handler Class, so we can set up a different frequency for each
             // State notification...
-
             if (stateRefreshFrequencies.size() > 0) {
                 this.executor = ThreadUtils.getSingleThreadScheduledExecutorService(id + "-HandlerStatusRefresh");
                 for (Handler handler : handlers.values()) {
@@ -166,10 +257,14 @@ public class P2P extends Thread {
             }
             logger.info("Thread Pool used: " + (runtimeConfig.useCachedThreadPoolForP2P()? "Cached" : "Fixed") + ", MaxThreads: " + runtimeConfig.getMaxNumThreadsForP2P());
 
-            startedLatch.countDown();   // the P2P is started
+            startedLatch.countDown();   // signal that the P2P is started
 
             // main processing, to be extended
-            stopLatch.await();
+            do {
+                eventsToProcess.acquire();
+                handlePendingToCloseConnections();
+                handlePendingToOpenConnections();
+            } while (stopLatch.getCount() != 0);
 
             // shutdown
             logger.info("Stopping ...");
@@ -212,6 +307,7 @@ public class P2P extends Thread {
     /** Initiate a graceful shutdown of the P2P. */
     public void initiateStop() {
         stopLatch.countDown();
+        eventsToProcess.release();
     }
 
     /**
@@ -268,13 +364,157 @@ public class P2P extends Thread {
         logger.info("Handler " + handlerId + " Config Updated.");
     }
 
+
+    // We register this class to LISTEN for some events that might be published by other handlers.
+    private void registerForEvents() {
+        eventBus.subscribe(DisconnectPeerRequest.class, e -> onDisconnectPeerRequest((DisconnectPeerRequest) e));
+        eventBus.subscribe(ConnectPeerRequest.class, e -> onConnectPeerRequest((ConnectPeerRequest) e));
+        eventBus.subscribe(ConnectPeersRequest.class, e -> onConnectPeersRequest((ConnectPeersRequest) e));
+        eventBus.subscribe(PeersBlacklistedEvent.class, e -> onPeersBlacklisted((PeersBlacklistedEvent) e));
+        eventBus.subscribe(PeersRemovedFromBlacklistEvent.class, e -> onPeersRemovedFromBlacklist((PeersRemovedFromBlacklistEvent) e));
+        eventBus.subscribe(ResumeConnectingRequest.class, e -> onResumeConnecting((ResumeConnectingRequest) e));
+        eventBus.subscribe(StopConnectingRequest.class, e -> onStopConnecting((StopConnectingRequest) e));
+        eventBus.subscribe(DisconnectPeersRequest.class, e -> onDisconnectPeers((DisconnectPeersRequest) e));
+    }
+
+    // Event Handlers:
+    private void onDisconnectPeerRequest(DisconnectPeerRequest request) {
+        logger.info("DisconnectPeerRequest received: " + request.getPeerAddress());
+        this.disconnect(request.getPeerAddress());
+    }
+
+    private void onConnectPeerRequest(ConnectPeerRequest request)        { this.connect(request.getPeerAddres()); }
+    private void onConnectPeersRequest(ConnectPeersRequest request)      { this.connect(request.getPeerAddressList()); }
+    private void onPeersBlacklisted(PeersBlacklistedEvent event)         { this.blacklist(event.getInetAddresses());}
+    private void onPeersRemovedFromBlacklist(PeersRemovedFromBlacklistEvent event)
+    { this.removeFromBlacklist(event.getInetAddresses());}
+
+    private void onResumeConnecting(ResumeConnectingRequest request) {
+//        if (super.state().equals(Service.State.STARTING) || super.state().equals(Service.State.STOPPING)) return;
+//        resumeConnecting();
+    }
+
+    private void onStopConnecting(StopConnectingRequest request) {
+//        if (super.state().equals(Service.State.STARTING) || super.state().equals(Service.State.STOPPING)) return;
+//        stopConnecting();
+    }
+
+    private void onDisconnectPeers(DisconnectPeersRequest request) {
+        if (request.getPeersToDisconnect() != null) {
+            this.disconnect(request.getPeersToDisconnect());
+        }
+        if (request.getPeersToKeep() != null) {
+            this.disconnectAllExcept(request.getPeersToKeep());
+        }
+    }
+
+    private void processDisconnectRequest(DisconnectPeerRequest request) {
+        processDisconnectRequests(Collections.singletonList(request));
+    }
+
+    private void processDisconnectRequests(List<DisconnectPeerRequest> requests) {
+        if (requests == null) return;
+        List<DisconnectPeerRequest> newList = requests.stream()
+                .filter(r -> !pendingToCloseConns.contains( new PeerAddress2DisconnectPeerRequest_Comparator(r.getPeerAddress()) ))
+                .filter(r -> (activeConns.containsKey(r.getPeerAddress()) || pendingToOpenConns.contains(r.getPeerAddress())))
+                .toList();
+        if (newList.size() > 0) {
+            logger.trace("{} : Registering " + newList.size() + " Peers for Disconnection...", this.id);
+            pendingToCloseConns.addAll(newList);
+            pendingToOpenConns.removeAll(newList.stream().map(DisconnectPeerRequest::getPeerAddress).toList());
+            eventsToProcess.release();
+        }
+    }
+
+    private void disconnectAllExcept(List<PeerAddress> peerAddresses) {
+        Predicate<PeerAddress> notToRemove = p -> !peerAddresses.contains(p);
+        List<DisconnectPeerRequest> activePeersToRemove = activeConns.keySet()
+                .stream()
+                .filter(notToRemove)
+                .map(p -> new DisconnectPeerRequest(p, PeerDisconnectedEvent.DisconnectedReason.DISCONNECTED_BY_LOCAL))
+                .collect(Collectors.toList());
+        processDisconnectRequests(activePeersToRemove);
+    }
+
+    /**
+     * Handle the pending connections to open. For each PeerAddress, it tries to open a Socket channel. If
+     * there is a limit in the maximum number of connections and we reach it, it does nothing.
+     */
+    private void handlePendingToOpenConnections() {
+        // loop over the pending Connections...
+        while (true) {
+            if (activeConns.size() >= networkConfig.getMaxSocketConnections()) break;
+
+            PeerAddress peerAddress = this.pendingToOpenConns.poll();
+            if (peerAddress == null) break;     // No more pending connections to open.
+
+            if (!shouldProcessPeer(peerAddress)) {
+                continue;
+            }
+
+            logger.trace("{} : {} : opening connection. Still pendingToOpen in Queue: {}, Active: {} ", this.id, peerAddress, this.pendingToOpenConns.size(), this.activeConns.size());
+            openConnection(peerAddress);
+        }
+    }
+
+    private boolean shouldProcessPeer(PeerAddress peerAddress) {
+        boolean processThisPeer;
+
+        processThisPeer = (peerAddress != null);
+        processThisPeer &= (!activeConns.containsKey(peerAddress));
+        processThisPeer &= (!blacklist.contains(peerAddress.getIp()));
+
+        return processThisPeer;
+    }
+
+    /**
+     * Find a NetworkController to open and handle the connection.
+     */
+    private void openConnection(PeerAddress peerAddress) {
+        // at the moment we only have one NetworkController
+        networkController.openConnection(peerAddress);
+        activeConns.put(peerAddress, networkController);
+    }
+
+    /**
+     * Handles the connections pending to close.
+     */
+    private void handlePendingToCloseConnections() {
+        DisconnectPeerRequest disconnectRequest;
+        while ((disconnectRequest = pendingToCloseConns.poll()) != null) {
+            PeerAddress peerAddress = disconnectRequest.getPeerAddress();
+            NetworkController networkController = activeConns.get(peerAddress);
+            if (networkController != null) {
+                logger.trace("{} : {} : Closing connection...", this.id, peerAddress);
+                networkController.closeConnection(peerAddress, disconnectRequest.getReason());
+                activeConns.remove(peerAddress);
+            }
+        }
+    }
+
+    /**
+     * Helper used to search if pendingToCloseConns contains a disconnect request for a specific peer
+     */
+    private static class PeerAddress2DisconnectPeerRequest_Comparator {
+        private final PeerAddress peerAddress;
+        public PeerAddress2DisconnectPeerRequest_Comparator(PeerAddress peerAddress) {
+            this.peerAddress = peerAddress;
+        }
+        @Override
+        public boolean equals(final Object o) {
+            final var other = (DisconnectPeerRequest) o;
+            assert other!=null; // should never be called with any other type of object
+            return peerAddress.equals(other.getPeerAddress());
+        }
+    }
+
     // Convenience method to get a reference to a P2PBuilder
     public static P2PBuilder builder(String id) {
         return new P2PBuilder(id);
     }
 
     public RuntimeConfig getRuntimeConfig()     { return this.runtimeConfig; }
-    public P2PConfig getNetworkConfig()     { return this.networkConfig; }
+    public P2PConfig getNetworkConfig()         { return this.networkConfig; }
     public ProtocolConfig getProtocolConfig()   { return this.protocolConfig; }
     public EventBus getEventBus()               { return this.eventBus;}
 }
